Update QDR for external IP and configuration
Updates the QDR configuration so that we don't create a loop (can't connect the two
central QDR routers together, only single direction). Updated the first QDR router to
provide a failover list of the next set of QDR routers (only 1 in this case).

Control the external IP correctly for the two external messaging services.